Move split2 method to table.rs
